requirements.txt
Flask
numpy
tensorflow
requests
google-cloud-firestore
scikit-learn
